+++ src/silfont/__init__.py
@@ -2,4 +2,4 @@
 __url__ = 'https://github.com/silnrsi/pysilfont'
 __copyright__ = 'Copyright (c) 2014-2024 SIL International (https://www.sil.org)'
 __license__ = 'Released under the MIT License (https://opensource.org/licenses/MIT)'
-__version__ = '1.8.1.dev4'
+__version__ = '1.8.1.dev5'