+++ CHANGELOG.md
@@ -14,9 +14,6 @@ General updates for new year, 2024
 
 ### Changed
 
-- Added --compregex to psfufo2ttf
-- Fix r-string SyntaxWarnings
-- Deprecate psfmakefea in favor of makefea
 - Handle CJK variants of ASCII characters in FTML builder that have
 the same base glyph name (that is, with no extension) as the ASCII base glyph name. (1.8.1.dev2)
 - Updated ttfchecks.py to match changes in Font Bakery checks (1.8.1.dev3)
@@ -25,6 +22,10 @@ the same base glyph name (that is, with no extension) as the ASCII base glyph na
 - Updated profiles and checks in fbtests/ to work with latest refactored Font Bakery v0.12.X
 - Fixed #86 "psfmakefea fails when there are different substitution types" (1.8.1.dev4)
 - Updated psfrunfbchecks to output a warning indicating the script is no longer to be used (1.8.1.dev4)
+- Deprecate psfmakefea in favor of makefea
+- Fix r-string SyntaxWarnings
+- Added --compregex to psfufo2ttf
+- psfexportanchors: Support public.skipExportGlyphs; add --ignoreglyphs option; correct/enhance documentation (1.8.1.dev5)
 
 ### Removed